Fold address space management into RPC core
