+++ cuda_server.cpp
@@ -37,7 +37,6 @@
 
 #include "cuda_compat.h"
 
-#include "address_space.h"
 #include "cache.h"
 #include "codegen/gen_rpc_ids.h"
 #include "copy_pipeline.h"
