+++ h2.cpp
@@ -1,4 +1,3 @@
-#include "address_space.h"
 #include "lupine_log.h"
 #include "rpc.h"
