+++ test/test_address_space.cpp
@@ -1,4 +1,3 @@
-#include "address_space.h"
 #include "rpc.h"
 
 #include <cerrno>
