+++ transport.cpp
@@ -1,6 +1,5 @@
 #include "transport.h"
 
-#include "address_space.h"
 #include "lupine_log.h"
 
 #include <algorithm>
